requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
Flask-Session==0.5.0
Flask-Login==0.6.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
